requirements.txt
accelerate==1.7.0
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorlog==6.9.0
einops==0.8.1
exceptiongroup==1.3.0
fastapi==0.115.12
ffmpeg-python==0.2.0
filelock==3.13.1
fsspec==2024.6.1
future==1.0.0
h11==0.16.0
hf-xet==1.1.3
huggingface-hub==0.32.4
idna==3.10
imageio==2.37.0
Jinja2==3.1.4
lazy_loader==0.4
Mako==1.3.10
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.11.0.86
packaging==25.0
pillow==11.0.0
platformdirs==4.3.8
psutil==7.0.0
pycuda==2025.1.1
pydantic==2.11.5
pydantic_core==2.33.2
python-multipart==0.0.20
pytools==2025.1.6
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scenedetect==0.6.6
scikit-image==0.25.2
scipy==1.15.3
sentencepiece==0.2.0
siphash24==1.7
sniffio==1.3.1
starlette==0.46.2
starlette-context==0.4.0
sympy==1.13.1
tifffile==2025.5.10
timm==1.0.15
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.52.4
triton==3.2.0
typing-inspection==0.4.1
typing_extensions==4.12.2
urllib3==2.4.0
uuid==1.30
uvicorn==0.34.3
python-dotenv==1.1.1
openai==1.100.2
sentence-transformers==5.1.0
python-dotenv==1.1.1